CAT-113 adds more text to the titels and cards in German
