+++ src/main/webapp/dashboard/i18n/i18n.properties
@@ -86,9 +86,6 @@ AllValid=Alle g\u00FCltig
 AllOnline=Alle Online
 NoData=Daten werden geladen...
 
-# Dashboard.view.xml
-connectorAutomate=Konnektor Automatisierung
-
 #Detail.controller.js
 msgSelectOneConnector=Bitte w\u00E4hlen Sie einen Konnektor aus.
 msgSelectOnlyOneConnector=Bitte w\u00E4hlen Sie nur einen Konnektor aus.
@@ -99,6 +96,8 @@ msgEditError=Fehler beim \u00E4ndern der Daten.
 SecureCommunicationUsingSSL/TLS=Kommunikation mittels SSL/TLS absichern
 CertificateBasedAuthentication=Zertifikatsbasierte Authentifizierung
 
-# EnabledAvailable.view.xml
+# AutomaticMonitoring.view.xml
 enabled=Aktiviert
 available=Verf\u00FCgbar
+connectorAutomate=Automatische Konnektor\u00FCberwachung
+introText=Um die automatischen \u00DCberwachungsprozesse zu starten, ziehen Sie bitte das entsprechende Element von der Sektion "Verf\u00FCgbar" in die Sektion "Aktiviert"
